fix(neutts): pad codes to enumerated shapes, switch to int64 dtype, cap at 1500
- CRITICAL: Pad speech_codes to next enumerated CoreML shape before decode.
  CoreML rejects predictions with shapes not in EnumeratedShapes list.
  Trim output audio proportionally to remove padding artifacts.
- MEDIUM: Switch from int32 to int64 throughout the pipeline (trace, export,
  MLMultiArray, speech_codes vector) to match original ONNX model input spec.
- LOW: Cap speech_codes accumulation at 1500 to guarantee max enumerated size.

+++ scripts/convert_neucodec_to_coreml.py
@@ -216,7 +216,7 @@ def main():
     TRACE_SIZE = 256
 
     print(f"[neucodec-coreml] Tracing PyTorch model (T={TRACE_SIZE})...", flush=True)
-    example = torch.zeros(1, 1, TRACE_SIZE, dtype=torch.int32)
+    example = torch.zeros(1, 1, TRACE_SIZE, dtype=torch.int64)
     with torch.no_grad():
         traced = torch.jit.trace(torch_model, example, check_trace=False)
 
@@ -231,7 +231,7 @@ def main():
             ct.TensorType(
                 name="codes",
                 shape=enumerated,
-                dtype=np.int32,
+                dtype=np.int64,
             )
         ],
         outputs=[ct.TensorType(name="audio", dtype=np.float32)],
